DTrace build fix
During config step just being "contented" by generating the header,
there is no symbols to attach for, no chance to work as is.
Changing a probe signature, on some platforms, pthread_t is an
opaque type thus casting to a type large enough to hold it
for all oses.

+++ memcached_dtrace.d
@@ -57,7 +57,7 @@ provider memcached {
     * @param connid the connection id
     * @param threadid the thread id
     */
-   probe conn__dispatch(int connid, int threadid);
+   probe conn__dispatch(int connid, int64_t threadid);
 
    /**
     * Allocate memory from the slab allocator.
